requirements.txt
requests==2.7.0
websocket-client==0.32.0
